+++ src/main/java/com/squid/config/SquidRestConfiguration.java
@@ -1,4 +1,4 @@
-package com.squid;
+package com.squid.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -7,7 +7,7 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurerAdapter;
 
 @Configuration
-public class SquidConfiguration {
+public class SquidRestConfiguration {
     @Bean
     public WebMvcConfigurer corsConfigurer() {
